Don't rely on unaligned accesses or little-endianness
This uses memcpy to avoid the unaligned accesses.  gcc on x86 seems to
do a reasonable job of optimizing them away.  gcc on ARM leaves some
memcpy calls in the generated code.

It might be better in terms of the code generated to use an approach
based on packaged structs (that's wwhat the Linux kernel does).  But
that would depend on compiler-specific syntax.