Makefile.am: Support building in a separate build dir & static linking
Building rabbitmq-c involves building sources too, which are then to
be used by other parts of the code. These built sources end up being
placed under the build directory, not the source. For this reason,
lets update the Makefile.am, so that AM_CFLAGS includes both
$(top_srcdir)/librabbitmq and $(top_builddir)/librabbitmq. Then,
everything that sets any kind of CFLAGS, also has to be updated to
include $(AM_CFLAGS), as to not override it.

While there, also shuffle around a few LDADDs, because with static
linking, librabbitmq.la needs to come last, so that the symbols
tools/libcommon.la uses from it will be found by the linker.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>