Report out of buf space error in amqp_table_encode
amqp_encode_8 and amqp_encode_bytes return 1 on success, 0 otherwise.
Correct the checking of the return value of these functions to match the
style used in amqp_table.c

Report AMQP_STATUS_INVALID_PARAMETER in the event an invalid field kind
is input (instead of abort()).

Report AMQP_STATUS_TABLE_TOO_BIG if the table is too large to be encoded
in the buffer provided.