Simplify the timer/timeout logic.
Instead of passing around a start time and timeout interval, calculate this
early and pass around the timeout value (as a implementation detail), then
calculate against current timestamp.  This simplifies the code, gives a more
realistic 'timeout' for the function at the cost of potentially a few more timer
pings.

Conflicts:
	librabbitmq/amqp_socket.c