build: Add check for posix_spawn functions
These are usually found in librt. On some systems clock_gettime is also
in librt so checking for it was implicitly satisfying the requirement
for the posix_spawn functions. On newer systems (glibc >2.17 for
example) clock_gettime is also found in libc so AC_SEARCH_LIBS() decides
that it is not necessary to add -lrt to LIBS.

Add an explicit check for posix_spawnp() to decide if -lrt is necessary.