Win32: use SRWLOCK instead of CRITIAL_SECTION
SRWLOCK allows for static-initialization of the lock, matching the
semantics of pthread_mutex that is used in OpenSSL bits, simplifying the
code somewhat.

Note that this API only became available in Windows Vista and newer.
Given Windows XP and Windows Server 2003 are no longer supported by
Microsoft, using this newer API is an acceptable tradeoff.