Lib: simplify socket connection logic
Refactor amqp_socket_open_inner so that the socket connection logic is
in it's own OS-specific socket_connect function, and the host resolution
and retry logic calls this. This makes the socket connection logic
easier to understand and reduces the number of win32 #ifdefs.

This also fixes an issue where multiple hostnames are not tried if
connect doesn't fail immediately.

Fixes #430