cmake: set default for BUILD_TOOLS to OFF
Then make BUILD_TOOLS_DOCS only usable when BUILD_TOOLS is ON and
explicitly specified.

This makes the project easier to add_submodule by not sometimes turning
on features based on the environment.  And makes it explict when things
are enabled vs something that is triggered based on the environment.

Signed-off-by: GitHub <noreply@github.com>