ci: use ninja to build under linux/macos
This is done to make the output when the build fails a bit easier to
read, specifically we'd like to see the compiler command used.

This theoretically will improve the build performance a bit by doing the
compilation in parallel, however the cost for installing ninja likely
more than offsets this.

Signed-off-by: GitHub <noreply@github.com>